#23 Clean up code -deleted unused files -removed commented out code -renamed some files -merged train and predict KNN into KNN_TP
