--- Code/Testing_Recognition.py
@@ -1,10 +1,0 @@
-from Sensor_Data_Reader import *
-from Feature_Extraction import *
-from Data_Frame_Merging import *
-from DF_to_CSV import *
-
-# extracted features
-testData = read_file('../DataSet/testing_gestures.csv')
-
-random_df = feature_extraction(testData, "random", "ring1")
-df_to_csv(random_df)
